Updating Dockerfile to remove dependency on hacked .s2i override
This project contains two build strategies, one based on a Dockerfile
and the other based on hijacking a S2I build from a parent image.

The latter seems to be an unnecessarily roundabout way to augment the
inheritted S2I behavior.  Its use of /tmp/.s2i/image_metadata.json to
override the LABELS on its source image do achieve a functional result,
but they do not cause the documentation metadata in some of those tags
to replace the documentation content from the base image, leading to a
lot of potential confusion.

Creating an S2I image builder is not a basic end-developer activity, so
it is plausible to accept that S2I image builds may presume to be
granted authorization to run as a Dockerfile based build, rather than
needing to be build by an S2I Builder themselves.